ci: enforce build+fmt+clippy+test on every PR; clean up warnings to green baseline
Expands the workflow to gate four checks on PRs:
  - cargo fmt --all -- --check (fails with a "run cargo fmt" message)
  - cargo clippy --workspace ... -- -D warnings
  - cargo build --workspace --all-targets
  - cargo test --workspace
All four use the same Tauri-crate excludes (cimmeria-app /
cimmeria-content-editor / cimmeria-scene-editor / sgw-launcher) and
hydrate external/recast at runtime to mirror what setup.ps1 does
locally.

Code-side changes to make `clippy -D warnings` green without
suppression:

- clippy.toml at the repo root raises two thresholds at the project
  level: `too-many-arguments-threshold = 14` (handlers in
  base/world_entry/methods/ take socket/connected/entity_to_addr/
  db_pool/cell_tx context — refactoring to context structs is a
  separate architectural change) and `type-complexity-threshold = 500`
  (the (Arc<UdpSocket>, Arc<Mutex<...>>, Arc<Mutex<...>>) tuple
  returned by test fixture builders is inherently nested).
- crates/defs/build/codegen.rs: emit #[derive(Default)] on generated
  *Props structs instead of a hand-written impl Default. Every
  property type already has a Default impl matching what the manual
  impl produced; this drops 14 derivable_impls warnings on the
  generated cimmeria-defs lib code.
- Idiom replacements (semantically identical, applied via
  cargo clippy --fix or by hand):
  * Option::map_or(false, f) → Option::is_some_and(f) (multiple sites)
  * (a + b - 1) / b → a.div_ceil(b) (upk/package.rs)
  * sort_by(|a, b| b.k.cmp(&a.k)) → sort_by_key(|x| Reverse(x.k))
    (chain.rs, several upk bins, build_package_index)
  * .min(N).max(M) → .clamp(M, N) (audit.rs, livewire.rs)
  * field assignment after Default::default() → struct-literal
    construction (mercury/nub.rs, upk/header.rs, auth/service.rs,
    base/mod.rs)
  * Vec::with_capacity + push push → vec! macro (mercury/protocol/session.rs)
  * if-else on is_none() → if let Some(x) = ... (cell/interactions/loot.rs;
    drops a fragile unwrap-after-is_none)
  * if e.is_player { 1025 } else { 1025 } → drop the dead branch
    (cell/abilities/damage_apply.rs)
- Doc-lint fixes:
  * /// → //! at the top of cell_methods/player/constants.rs
  * Reflowed the static_mesh.rs binary-layout doc as a code block so
    CommonMark sub-list indentation rules don't fight clippy
  * Removed a `+` that started a phantom list in mercury/aoi.rs
- Dead-code removal (no call sites, found via grep):
  * query_player_load_data_by_account in player_load/core.rs
  * clear_buyback_items + the duplicate INV_BUYBACK in vendor/store.rs
    (the canonical INV_BUYBACK lives in cimmeria-entity)
  * SGWGMPLAYER_CLASS_ID in mercury/mod.rs
  * Trimmed unread fields (type_id, slot_id, bound, durability,
    charges) from InventoryInstanceRow + narrowed the SELECT
- pub use re-exports in vendor/mod.rs, inventory/mod.rs,
  player_load/mod.rs trimmed to the items actually consumed downstream.
- One unused parameter: print_sequence_tree's `all_nodes` in
  upk/extract_kismet.rs was forwarded only to its own recursive call
  and never read; removed from signature + call sites.
- Useless test assertion: assert!(usize::len() >= 0) in
  defs/registry.rs → drop the assert, keep the construct-without-panic
  intent in a comment.
- Parens nit in cell/service/base_messages.rs.
- A 2-blank-line gap in cell_methods/player/dispatch.rs collapsed; the
  match arms there now resolve constants via `use super::constants::*`
  at the lib level instead of repeating `super::constants::FOO` in
  every arm (no semantic change).

CLAUDE.md: new "Pre-PR checklist" section with the exact commands.

Verified locally:
  - cargo clippy --workspace [excludes] --all-targets -- -D warnings
  - cargo fmt --all -- --check
  - DATABASE_URL=… cargo test -p cimmeria-services --lib
    -- --test-threads=1 — 357 / 357 pass

Behavior is unchanged; the live-DB test failures observed under heavy
parallelism are pre-existing sentinel-range collisions, not regressions
from this cleanup (every test passes individually and serialized).

+++ crates/services/src/cell/abilities/use_ability.rs
@@ -269,7 +269,7 @@ pub async fn handle_use_ability(
                 seq_args.extend_from_slice(&0.0f32.to_le_bytes()); // ImpactTime
                 seq_args.extend_from_slice(&0u32.to_le_bytes()); // NameValuePairs array count = 0
                 seq_args.push(0); // ViewType = KISMET_VIEW_Witness
-                seq_args.extend_from_slice(&(effect_seq as i32).to_le_bytes()); // InstanceId
+                seq_args.extend_from_slice(&effect_seq.to_le_bytes()); // InstanceId
                 send_entity_method(entity_id, 1, seq_args, tx, space_mgr).await;
                 // 1 = onSequence
             }
@@ -288,7 +288,7 @@ pub async fn handle_use_ability(
             seq_args.extend_from_slice(&0.0f32.to_le_bytes()); // ImpactTime
             seq_args.extend_from_slice(&0u32.to_le_bytes()); // NameValuePairs array count = 0
             seq_args.push(0); // ViewType = KISMET_VIEW_Witness
-            seq_args.extend_from_slice(&(effect_seq as i32).to_le_bytes()); // InstanceId
+            seq_args.extend_from_slice(&effect_seq.to_le_bytes()); // InstanceId
             send_entity_method(entity_id, 1, seq_args, tx, space_mgr).await; // 1 = onSequence
         } else {
             tracing::debug!(
